Added Version 1.7.1 (Scripts) 03-27-25

+++ scripts/extension_systems/dialogue/tag_query_database.lua
@@ -95,6 +95,8 @@ TagQueryDatabase.define_rule = function (self, rule_definition)
 
 	rule_definition.n_criterias = num_criterias
 
+	local add_rule = true
+
 	for i = 1, num_criterias do
 		local criteria = criterias[i]
 		local context_name = criteria[1]
@@ -130,24 +132,29 @@ TagQueryDatabase.define_rule = function (self, rule_definition)
 			criteria[4] = value
 		elseif value_type == "table" then
 			criteria[4] = value
+			add_rule = (not table.is_empty(value) or false) and add_rule
 		end
 	end
 
-	local rule_id = RuleDatabase.add_rule(self._database, dialogue_name, num_criterias, criterias)
+	if add_rule then
+		local rule_id = RuleDatabase.add_rule(self._database, dialogue_name, num_criterias, criterias)
 
-	self._rule_id_mapping[rule_id] = rule_definition
-	self._rule_id_mapping[rule_definition.name] = rule_id
-	self._rules_n = self._rules_n + 1
+		self._rule_id_mapping[rule_id] = rule_definition
+		self._rule_id_mapping[rule_definition.name] = rule_id
+		self._rules_n = self._rules_n + 1
+	end
 end
 
 TagQueryDatabase.remove_rule = function (self, rule_name)
 	local rule_id = self._rule_id_mapping[rule_name]
 
-	self._rule_id_mapping[rule_id] = nil
-	self._rule_id_mapping[rule_name] = nil
-	self._rules_n = self._rules_n - 1
+	if rule_id then
+		self._rule_id_mapping[rule_id] = nil
+		self._rule_id_mapping[rule_name] = nil
+		self._rules_n = self._rules_n - 1
 
-	RuleDatabase.remove_rule(self._database, rule_id)
+		RuleDatabase.remove_rule(self._database, rule_id)
+	end
 end
 
 TagQueryDatabase.num_rules = function (self)
